+++ docs/development/releasenotes.rst
@@ -26,3 +26,16 @@ Release Notes
 12. *Templates*. CSS updates to the Tufte book template.
 13. *Utils*. Add general utilities for classes, dicts, lists and strings.
 14. *Tests*. Updates to tests to work with new APIs and refactorings.
+15. *TagFactory*. Refactor into a submodule and separate documentation
+16. *Substitution*. Create a new Substitution tag for placing the contents of a
+    macro in a tag.
+17. *AST*. Make the parsing of tags with content optional.
+18. *Context*. Make the context weakref-able so that tags only need to store a
+    weakref to contexts.
+19. *Context*. Added a utility function for setting the 'content' attribute of
+    asts and other objects.
+20. *LabelManager*. Converted the root_context attribute to a weakref
+21. *Tag*. Converted the context attribute to a weakref
+22. *Tag utils*. Added a repl_tags utility function for changing all tags of a
+    specified class in an AST element or tag to a replacement string. Included
+    tests.